requirements.txt
aiohttp==3.8.1
awesome-slugify==1.6.5
certifi==2020.6.20
chardet==3.0.4
gql==3.2.0
idna==2.10
python-gitlab==2.5.0
pytz==2022.1
pyyaml==6.0.0
regex==2020.10.28
requests==2.24.0
requests-toolbelt==0.9.1
sentry-sdk==0.19.5
Unidecode==0.4.21
urllib3==1.25.11
